backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# Celery & Task Queue
celery==5.3.4
redis==5.0.1
flower==2.0.1

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.2
lightgbm==4.1.0
catboost==1.2.2
optuna==3.4.0

# XAI Libraries
shap==0.43.0
lime==0.2.0.1
quantus==0.5.2

# Data Processing
pandas==2.1.3
numpy==1.26.2
pyarrow==14.0.1

# Kaggle Integration
kaggle==1.5.16

# Storage
supabase==2.0.3

# Logging & Monitoring
structlog==23.2.0
python-json-logger==2.0.7

# HTTP Client
httpx==0.24.1
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
